{{cookiecutter.directory_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = ["BACKBONE <office@backbone.sk>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.0"
django_api_forms = "1.0.0rc7"
python-dotenv = "^0.21.0"
django-filter = "^22.1"
argon2-cffi = "^21.3.0"
sentry-sdk = "^1.11"
psycopg2-binary = "^2.9.5"
porcupine-python = "^0"
django-imap-backend = "^0"
redis = "^4.4"

[tool.poetry.dev-dependencies]
fabric = "^2.7.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
